+++ aioesphomeapi/model.py
@@ -542,8 +542,8 @@ class ClimateInfo(EntityInfo):
     )
     supports_current_humidity: bool = False
     supports_target_humidity: bool = False
-    visual_min_humidity: int = 0
-    visual_max_humidity: int = 0
+    visual_min_humidity: float = 0
+    visual_max_humidity: float = 0
     supports_aux_heat: bool = False
 
     def supported_presets_compat(self, api_version: APIVersion) -> list[ClimatePreset]:
@@ -588,8 +588,8 @@ class ClimateState(EntityState):
         default=ClimatePreset.NONE, converter=ClimatePreset.convert
     )
     custom_preset: str = ""
-    current_humidity: int = 0
-    target_humidity: int = 0
+    current_humidity: float = 0
+    target_humidity: float = 0
     aux_heat: bool = False
 
     def preset_compat(self, api_version: APIVersion) -> ClimatePreset | None:
